code restructure: relocate eta files
