+++ src/components/eta/BookmarkCardEtaUnified.tsx
@@ -2,9 +2,9 @@ import { Badge, Text } from "@fluentui/react-components";
 import { Card, CardHeader } from "@fluentui/react-components/unstable";
 import { Link } from "react-router-dom";
 
-import { CountdownSec } from "../../components/countdown/CountdownSec";
 import { stopBookmarkWithEta } from "../../models/etaObjects";
 import { fluentStyles } from "../../styles/fluent";
+import { CountdownSec } from "../countdown/CountdownSec";
 
 export function BookmarkCardEtaUnified(props: { item: stopBookmarkWithEta }) {
   const fluentStyle = fluentStyles();
