+++ src/routes/Home.tsx
@@ -2,7 +2,7 @@ import { Button, Input } from "@fluentui/react-components";
 import { SetStateAction, useCallback, useState } from "react";
 import { useTranslation } from "react-i18next";
 
-import FavouriteEta from "../features/eta/FavouriteEta";
+import FavouriteEta from "../components/eta/FavouriteEta";
 import { fluentStyles } from "../styles/fluent";
 import useNavigate from "./navigate";
